feat: Add sidebar
