+++ package.json
@@ -29,6 +29,7 @@
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^3.0.0",
+    "tailwindcss": "^3.2.7",
     "typescript": "^4.6.4",
     "vite": "^4.0.0"
   }
